Ignore *.swp files
This commit updates the .gitignore file to include *.swp files, which are temporary files created by various editors. Additionally, the commit removes the commented out section for ignoring the entire IDEA folder. This section is not recommended, and it's better to use a .gitignore file to ignore specific files or folders. The README_tools_template.md file remains unchanged.